+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 
 - add assetTransactionType.COAReserveTransaction
+- add project.hasMerchantReturnPolicy.identifier
+- add seller.hasMerchantReturnPolicy.identifier
 
 ### Changed